requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlmodel==0.0.22
pydantic[email]==2.9.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
jinja2==3.1.4
faker==30.1.0
httpx==0.27.2
python-dotenv==1.0.1
starlette<0.46